AI-service/pyproject.toml
[project]
name = "wellness-llm-training"
version = "0.1.0"
description = "Dataset preparation pipeline for wellness LLM fine-tuning"
requires-python = ">=3.10"
dependencies = [
    "datasets>=2.14.0",
    "pyyaml>=6.0",
    "jsonlines>=4.0.0",
    "tqdm>=4.65.0",
]

[tool.setuptools.packages.find]
where = ["scripts"]

[project.optional-dependencies]
dev = ["ruff", "mypy"]
